requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
cloudinary==1.44.0
crispy-bootstrap5==2024.10
Django==4.2.20
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-nested-admin==4.1.1
django-taggit==6.1.0
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
fontawesomefree==6.6.0
gunicorn==23.0.0
idna==3.6
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
python-dotenv==1.1.0
python-monkey-business==1.1.0
reportlab==4.4.0
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.4.0
whitenoise==6.8.2
